+++ layers/WPSchema/packages/blocks/src/FilterInputs/BlockTypeAttributeAutoGenerateControlFilterInput.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PoPWPSchema\Blocks\FilterInputs;
+
+use PoP\ComponentModel\FilterInputs\AbstractValueToQueryFilterInput;
+
+class BlockTypeAttributeAutoGenerateControlFilterInput extends AbstractValueToQueryFilterInput
+{
+    protected function getQueryArgKey(): string
+    {
+        return 'autoGenerateControl';
+    }
+}